% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{Pacote Qualquer Coisa}
\usage{
power(exp)
}
\arguments{
\item{exp}{Variável numérica que irá exponencializar \code{x}}
}
\description{
Esse pacote faz qualquer coisa. Basicamente estamos estudando
a consturção de pacotes em R. As funções aqui presentes não necessariamente
estão relacionadas.
}
\details{
Outro detalhe qualquer que seja do seu interesse você colocará aqui.
}
\note{
Bla bla bla bla
}
\examples{
quadrado <- power(2)
quadrado(3)
}
\author{
Pedro Rafael D. Marinho (\email{pedro.rafael.mariho@gmail.com})
}
