% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{roubando_lillie_test}
\alias{roubando_lillie_test}
\title{Função para teste de normalidade em amostras pequenas}
\usage{
roubando_lillie_test(x)
}
\arguments{
\item{x}{vetor atômico para se testar a normalidade}
}
\description{
Função para teste de normalidade em amostras pequenas
}
\details{
Essa função é apenas um envólucro da função \code{\link{lillie.test}}
do pacote \strong{nortest}
}
